cityService/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.1
schedule==1.2.1
psutil==5.9.0
prometheus-client==0.19.0
prometheus-flask-exporter==0.23.0
PyJWT==2.8.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.1
kafka-python==2.0.2
confluent-kafka==2.3.0